refactor(game-engine): extract handleBlock + import cleanup (S27.8.8)
Slice 8 de S27.8 (refactor monolith `actions.ts`).

Extraction (`actions/block-action.ts`)
- `handleBlock` (~180l) : flow complet de blocage standard / blitz
  pendant un mouvement. Couvre Dump-off check, Foul Appearance,
  calcul de force avec modifiers ST registry-based, Dauntless,
  resolution 1 de immediate ou pose pendingBlock pour
  resolution differee par `handleBlockChoose`.
- Aucune dependance interne au dispatcher (handleDumpOffChoose /
  handleMultiBlock / resolveFrenzyBlock / resolveMultipleBlock
  appellent handleBlock — la dependance est unidirectionnelle, pas
  cyclique).

Imports orphelins retires de `actions.ts`
- `BlockResult`, `ActionType` (types — utilises uniquement par les
  handlers extraits).
- `getPickupSkillModifiers`, `canSkillReroll` (skill-bridge).
- `collectModifiers` (skill-registry).
- `calculatePickupModifiers` (movement).
- `performDodgeRoll`, `performPickupRoll`, `performArmorRoll`,
  `rollBlockDice`, `blockResultFromRoll` (utils/dice).
- `performPickupRollWithNotification`,
  `rollBlockDiceWithNotification`,
  `rollBlockDiceManyWithNotification` (utils/dice-notifications).
- `resolveBlockResult` (mechanics/blocking).
- `getPlayerAction`, `canUseTeamReroll` (core/game-state).

Resultat
- `actions.ts` 1517 -> 1329 lignes (-188 nettes ; cumul -1377 vs
  origine 2706 = ~50.9%).
- `block-action.ts` ajoute a la liste d'exclusion coverage.
- 4888 tests game-engine passent. Coverage 25.33% maintenue.
- Lint warnings 1038 -> 1029 (cleanup des imports orphelins).

Reste pour atteindre la cible <= 600 lignes :
- handleBlitz (~230l)
- handleDodge (~120l)
- handleMultiBlock + handleDumpOffChoose + resolveFrenzyBlock +
  resolveMultipleBlock
- handleLeap (depend de getLegalMoves, cycle a casser)
- handleMove
- applyMove dispatcher central
